requirement.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
h11==0.12.0
idna==2.10
lxml==4.6.2
psycopg2-binary==2.8.6
pydantic==1.7.3
requests==2.25.1
soupsieve==2.1
starlette==0.13.6
typing-extensions==3.7.4.3
urllib3==1.26.2
uvicorn==0.13.3
